codi-app/requirements.txt
pydbus
pyserial
evdev
psutil
